fix: race condition while publishing wm inside reduce
