+++ pkg/reduce/data_forward_test.go
@@ -410,7 +410,7 @@ func TestDataForward_StartWithNoOpWM(t *testing.T) {
 
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(child, keyedVertex, CounterReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publisher, idleManager, windower)
+	op := pnf.NewProcessAndForward(child, keyedVertex, CounterReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publisher, idleManager, windower)
 
 	var reduceDataForwarder *DataForward
 	reduceDataForwarder, err = NewDataForward(child, keyedVertex, fromBuffer, toBuffer, pbqManager, storeManager, CounterReduceTest{}, wmpublisher, publisher,
@@ -507,7 +507,7 @@ func TestReduceDataForward_IdleWM(t *testing.T) {
 	windower := fixed.NewWindower(5*time.Second, keyedVertex)
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffers))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, CounterReduceTest{}, toBuffers, pbqManager, CounterReduceTest{}, publisherMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, CounterReduceTest{}, toBuffers, pbqManager, CounterReduceTest{}, publisherMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	reduceDataForward, err = NewDataForward(ctx, keyedVertex, fromBuffer, toBuffers, pbqManager, storeManager, CounterReduceTest{}, f, publisherMap,
@@ -717,7 +717,7 @@ func TestReduceDataForward_Count(t *testing.T) {
 	windower := fixed.NewWindower(60*time.Second, keyedVertex)
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, CounterReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publisherMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, CounterReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publisherMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	reduceDataForward, err = NewDataForward(ctx, keyedVertex, fromBuffer, toBuffer, pbqManager, storeManager, CounterReduceTest{}, f, publisherMap,
@@ -803,7 +803,7 @@ func TestReduceDataForward_AllowedLatencyCount(t *testing.T) {
 
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, CounterReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publisherMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, CounterReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publisherMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	allowedLatency := 1000
@@ -891,7 +891,7 @@ func TestReduceDataForward_Sum(t *testing.T) {
 	windower := fixed.NewWindower(2*time.Minute, keyedVertex)
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	reduceDataForward, err = NewDataForward(ctx, keyedVertex, fromBuffer, toBuffer, pbqManager, storeManager, CounterReduceTest{}, f, publishersMap,
@@ -977,7 +977,7 @@ func TestReduceDataForward_Max(t *testing.T) {
 	windower := fixed.NewWindower(5*time.Minute, keyedVertex)
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, MaxReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, MaxReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	reduceDataForward, err = NewDataForward(ctx, keyedVertex, fromBuffer, toBuffer, pbqManager, storeManager, CounterReduceTest{}, f, publishersMap,
@@ -1064,7 +1064,7 @@ func TestReduceDataForward_FixedSumWithDifferentKeys(t *testing.T) {
 
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	reduceDataForward, err = NewDataForward(ctx, keyedVertex, fromBuffer, toBuffer, pbqManager, storeManager, CounterReduceTest{}, f, publishersMap,
@@ -1171,7 +1171,7 @@ func TestReduceDataForward_SumWithDifferentKeys(t *testing.T) {
 
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, SessionSumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, SessionSumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	reduceDataForward, err = NewDataForward(ctx, keyedVertex, fromBuffer, toBuffer, pbqManager, storeManager, SessionSumReduceTest{}, f, publishersMap,
@@ -1276,7 +1276,7 @@ func TestReduceDataForward_NonKeyed(t *testing.T) {
 
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	reduceDataForward, err = NewDataForward(ctx, nonKeyedVertex, fromBuffer, toBuffer, pbqManager, storeManager, CounterReduceTest{}, f, publishersMap,
@@ -1367,7 +1367,7 @@ func TestDataForward_WithContextClose(t *testing.T) {
 
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, CounterReduceTest{}, publishersMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	reduceDataForward, err = NewDataForward(cctx, keyedVertex, fromBuffer, toBuffer, pbqManager, storeManager, CounterReduceTest{}, f, publishersMap,
@@ -1465,7 +1465,7 @@ func TestReduceDataForward_SumMultiPartitions(t *testing.T) {
 
 	idleManager, err := wmb.NewIdleManager(1, len(toBuffer))
 	assert.NoError(t, err)
-	op := pnf.NewPnFManager(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, &myForwardTestRoundRobin{}, publishersMap, idleManager, windower)
+	op := pnf.NewProcessAndForward(ctx, keyedVertex, SumReduceTest{}, toBuffer, pbqManager, &myForwardTestRoundRobin{}, publishersMap, idleManager, windower)
 
 	var reduceDataForward *DataForward
 	reduceDataForward, err = NewDataForward(ctx, keyedVertex, fromBuffer, toBuffer, pbqManager, storeManager, &myForwardTestRoundRobin{}, f, publishersMap,
