+++ test/reduce-one-e2e/testdata/simple-keyed-reduce-pipeline.yaml
@@ -4,7 +4,7 @@ metadata:
   name: simple-sum
 spec:
   watermark:
-    maxDelay: 20s
+    maxDelay: 30s
   limits:
     readBatchSize: 50
   vertices:
