+++ test/reduce-one-e2e/testdata/simple-non-keyed-reduce-pipeline.yaml
@@ -4,7 +4,7 @@ metadata:
   name: reduce-sum
 spec:
   watermark:
-    maxDelay: 20s
+    maxDelay: 30s
   limits:
     readBatchSize: 50
   vertices:
