+++ test/reduce-one-e2e/testdata/simple-reduce-pipeline-wal.yaml
@@ -4,7 +4,7 @@ metadata:
   name: even-odd-sum
 spec:
   watermark:
-    maxDelay: 20s
+    maxDelay: 30s
   limits:
     readBatchSize: 50
   vertices:
